Fixed a bug related with blocking the users. The title of the block button and the string we were searching for was not equal so the code wasn't recognizing the blocked users.
