+++ app/src/main/res/layout/activity_browse_profiles.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".browseProfiles.BrowseProfilesActivity">
+    tools:context=".browseProfiles.BrowseProfilesUnblockedActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
