+++ app/src/main/res/layout/activity_browse_profiles_blocked.xml
@@ -6,6 +6,14 @@
     android:layout_height="match_parent"
     tools:context=".browseProfiles.BrowseProfilesBlockedActivity">
 
+    <include
+        android:id="@+id/toolbar"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_alignParentTop="true"
+        layout="@layout/toolbar"
+        />
+
     <TextView
         android:id="@+id/blockedMessage"
         android:layout_width="wrap_content"
